requirements.txt
ipython==7.34.0
pyserial==3.5
